Enable interrupts in the compartment switcher.
Preemption during a compartment switch shouldn't matter: the compartment
switching code touches only thread-local values (registers, the trusted
stack, and the stack).  It currently runs for a potentially long time
with interrupts disabled while it zeroes the stack.  I was going to make
it just enable interrupts for the zeroing loop, but I'm increasingly
confident that it is safe to enable them for the whole switcher path.